feat(backend): multi-store chain + 3-tier roles, drop sales-return flow
Đa cửa hàng (multi-store) cho toàn chuỗi:
- Thêm bảng/entity `stores`; gắn store_id vào users, goods_receipts, work_shifts,
  shelves, invoices (lô/dòng bán/chuyển-kệ thừa hưởng cửa hàng qua chứng từ).
- View tồn v_batch_stock/v_product_stock thêm chiều store_id (grain product×store).
- store_config 1–1 theo cửa hàng (ngân hàng/VietQR/WEB2M/Telegram riêng từng cửa hàng);
  telegram_recipients theo cửa hàng.
- Mọi truy vấn tổng hợp (dashboard/báo cáo/tồn/ABC) lọc theo cửa hàng; FIFO bán chỉ
  rút lô của đúng cửa hàng.

Phân quyền 3 tầng ADMIN > MANAGER > STAFF (Spring role hierarchy):
- ADMIN = quản trị TOÀN CHUỖI (không gắn cửa hàng): quản lý cửa hàng, tài khoản,
  catalog/giá/khuyến mãi/NCC (HQ-only), cấu hình từng cửa hàng, báo cáo gộp chuỗi.
- MANAGER/STAFF gắn 1 cửa hàng (StoreContext suy ra từ token); chặn truy cập chéo
  cửa hàng theo id (assertSameStore); nhân viên chỉ xem hoá đơn của ca mình.
- Cửa hàng đóng (INACTIVE) → nhân viên cửa hàng đó không đăng nhập được (token vô hiệu).

Bỏ HẲN chức năng trả hàng/hoàn tiền: xoá SalesReturn(+Item), repo, service, controller,
DTO, projection; viết lại view tồn (bỏ nhánh trả); doanh thu là GỘP (không trừ hoàn).

Migration idempotent cho CSDL cũ (backfill store_id, chuyển CASHIER→STAFF, CHAIN_ADMIN→ADMIN,
siết NOT NULL). 27/27 unit test PASS, compile sạch.

+++ backend/src/main/java/com/pos/repository/StoreConfigRepository.java
@@ -3,5 +3,6 @@ package com.pos.repository;
 import com.pos.entity.StoreConfig;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StoreConfigRepository extends JpaRepository<StoreConfig, Byte> {
+/** Cấu hình theo chi nhánh — khóa = stores.id (Long). */
+public interface StoreConfigRepository extends JpaRepository<StoreConfig, Long> {
 }
